% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{test_uniqueness}
\alias{test_uniqueness}
\title{Test uniqueness of data table}
\usage{
test_uniqueness(data_table, index_vars, print_pos = TRUE)
}
\arguments{
\item{data_table}{A data frame of data table of which uniqueness should
be tested.}

\item{index_vars}{Vector of strings, which specify the columns of
data_table according to which uniqueness should be tested
(e.g. country and year).}
}
\value{
TRUE if data_table is unique, FALSE and a warning if it is not.
}
\description{
Tests whether a data.table has unique rows.
}
\concept{update_dataset_helpers}
